+++ stroom-core-client-widget/src/main/java/stroom/widget/menu/client/presenter/IconParentMenuItem.java
@@ -35,21 +35,21 @@ public class IconParentMenuItem extends IconMenuItem implements HasChildren {
                                  final SvgImage disabledIcon,
                                  final IconColour iconColour,
                                  final SafeHtml text,
+                                 final SafeHtml tooltip,
                                  final Action action,
                                  final boolean enabled,
                                  final boolean highlight,
-                                 final String tooltip,
                                  final Future<List<Item>> children) {
         super(priority,
                 enabledIcon,
                 disabledIcon,
                 iconColour,
                 text,
+                tooltip,
                 action,
                 enabled,
                 null,
-                highlight,
-                tooltip);
+                highlight);
         this.children = children;
     }
 
@@ -89,10 +89,10 @@ public class IconParentMenuItem extends IconMenuItem implements HasChildren {
                     disabledIcon,
                     iconColour,
                     text,
+                    tooltip,
                     action,
                     enabled,
                     highlight,
-                    tooltip,
                     children);
         }
     }