+++ stroom-core-client-widget/src/main/java/stroom/widget/menu/client/presenter/MenuItem.java
@@ -25,17 +25,20 @@ import com.google.gwt.user.client.Command;
 public abstract class MenuItem extends Item {
 
     private final SafeHtml text;
+    private final SafeHtml tooltip;
     private final Action action;
     private final Command command;
     private final boolean enabled;
 
     protected MenuItem(final int priority,
                        final SafeHtml text,
+                       final SafeHtml tooltip,
                        final Action action,
                        final boolean enabled,
                        final Command command) {
         super(priority);
         this.text = text;
+        this.tooltip = tooltip;
         this.action = action;
         this.enabled = enabled;
         this.command = command;
@@ -43,6 +46,10 @@ public abstract class MenuItem extends Item {
 
     public SafeHtml getText() {
         return text;
+    }
+
+    public SafeHtml getTooltip() {
+        return tooltip;
     }
 
     public Action getAction() {
@@ -65,10 +72,10 @@ public abstract class MenuItem extends Item {
             extends Item.AbstractBuilder<T, B> {
 
         protected SafeHtml text;
+        protected SafeHtml tooltip;
         protected Action action;
         protected Command command;
         protected boolean enabled = true;
-        protected String tooltip;
 
         public B text(final SafeHtml text) {
             this.text = text;
@@ -77,6 +84,11 @@ public abstract class MenuItem extends Item {
 
         public B text(final String text) {
             this.text = SafeHtmlUtil.from(text);
+            return self();
+        }
+
+        public B tooltip(final String tooltip) {
+            this.tooltip = SafeHtmlUtil.from(tooltip);
             return self();
         }
 
@@ -92,11 +104,6 @@ public abstract class MenuItem extends Item {
 
         public B enabled(final boolean enabled) {
             this.enabled = enabled;
-            return self();
-        }
-
-        public B tooltip(final String tooltip) {
-            this.tooltip = tooltip;
             return self();
         }