+++ stroom-core-client-widget/src/main/java/stroom/widget/menu/client/presenter/SimpleMenuItem.java
@@ -25,10 +25,11 @@ public class SimpleMenuItem extends MenuItem {
 
     protected SimpleMenuItem(final int priority,
                              final SafeHtml text,
+                             final SafeHtml tooltip,
                              final Action action,
                              final boolean enabled,
                              final Command command) {
-        super(priority, text, action, enabled, command);
+        super(priority, text, tooltip, action, enabled, command);
     }
 
     public static class Builder extends AbstractBuilder<SimpleMenuItem, Builder> {
@@ -42,6 +43,7 @@ public class SimpleMenuItem extends MenuItem {
             return new SimpleMenuItem(
                     priority,
                     text,
+                    tooltip,
                     action,
                     enabled,
                     command);