+++ stroom-core-client/src/main/java/stroom/dashboard/client/embeddedquery/BasicEmbeddedQuerySettingsViewImpl.java
@@ -18,6 +18,7 @@ package stroom.dashboard.client.embeddedquery;
 
 import stroom.dashboard.client.embeddedquery.BasicEmbeddedQuerySettingsPresenter.BasicEmbeddedQuerySettingsView;
 import stroom.widget.tickbox.client.view.CustomCheckBox;
+import stroom.widget.valuespinner.client.ValueSpinner;
 
 import com.google.gwt.dom.client.Style.Unit;
 import com.google.gwt.uibinder.client.UiBinder;
@@ -48,10 +49,15 @@ public class BasicEmbeddedQuerySettingsViewImpl
     CustomCheckBox autoRefresh;
     @UiField
     TextBox refreshInterval;
+    @UiField
+    ValueSpinner pageSize;
 
     @Inject
     public BasicEmbeddedQuerySettingsViewImpl(final Binder binder) {
         widget = binder.createAndBindUi(this);
+        pageSize.setValue(100);
+        pageSize.setMin(1);
+        pageSize.setMax(10000);
     }
 
     @Override
@@ -115,6 +121,16 @@ public class BasicEmbeddedQuerySettingsViewImpl
     @Override
     public void setRefreshInterval(final String refreshInterval) {
         this.refreshInterval.setText(refreshInterval);
+    }
+
+    @Override
+    public int getPageSize() {
+        return pageSize.getIntValue();
+    }
+
+    @Override
+    public void setPageSize(final int pageSize) {
+        this.pageSize.setValue(pageSize);
     }
 
     public void onResize() {