+++ stroom-processor/stroom-processor-impl/src/main/java/stroom/processor/impl/ProcessorFilterServiceImpl.java
@@ -643,7 +643,7 @@ class ProcessorFilterServiceImpl implements ProcessorFilterService, HasUserDepen
                 })
                 .filter(Objects::nonNull)
 //                    .filter(userDependency ->
-//                            GwtNullSafe.getOrElse(
+//                            NullSafe.getOrElse(
 //                                    userDependency.getDocRef(),
 //                                    docRef -> securityContext.hasDocumentPermission(docRef, DocumentPermission.VIEW),
 //                                    true))